Create README.md
asxWACC is an R-package to calculate weighted average cost of capital (WACC) of ASX stocks using Yahoo Finance data. It can generate WACC of a firm while taking ticker as input. The expected market return is 6.5% as the average return over the last 10 years (2011-2020) for Australian Shares is 6.5%.

WACC = cost of debt x weight of debt x (1-tax_rate) + cost of equity x weight of equity